refactor(plugins): gather the retrieval helpers into one module
collect_document_ids, resolve_document_blocks and its chunk fetch are all one job: read the
document references out of a turn and put back what retrieval found. They were split across
messages.py and routes/utils, so following that job meant opening two files and skipping the
functions in between. They are now routes/utils/rag_search.py.

extract_query_text stays in messages.py. It is read before scope is judged, again when history is
deduplicated, and again when the messages are reassembled, so filing it under retrieval would put
the scope gate's own input behind a RAG import.

routes/utils/__init__.py is left with resolve_tools and parse_metadata_list — 41 lines, from 349
before this stack.
