+++ sparkth/plugins/chat/messages.py
@@ -2,14 +2,12 @@
 
 A turn arrives either as a plain string or as content blocks — text beside an uploaded document,
 or a reference to one already ingested — so nothing downstream can assume where the words are.
-These read one thing each out of that shape and are used by the route and by the stream alike,
-which is why they live here rather than under either.
+Reading the question back out is needed before scope is judged, before retrieval is considered,
+and again when the messages are reassembled, which is why it lives here and not under any one of
+those.
 """
 
-from sparkth.lib.log import get_logger
 from sparkth.plugins.chat.schemas import ChatMessage
-
-logger = get_logger(__name__)
 
 
 def extract_query_text(messages: list[ChatMessage]) -> str:
@@ -26,19 +24,3 @@ def extract_query_text(messages: list[ChatMessage]) -> str:
         if joined:
             return joined
     return ""
-
-
-def collect_document_ids(messages: list[ChatMessage]) -> list[int]:
-    document_ids: list[int] = []
-    for msg in messages:
-        if not isinstance(msg.content, list):
-            continue
-        for block in msg.content:
-            if not isinstance(block, dict) or block.get("type") != "drive_file":
-                continue
-            raw_id = block.get("file_id")
-            if raw_id is None:
-                logger.warning("Skipping document attachment block missing file_id in stream: %s", block)
-                continue
-            document_ids.append(int(raw_id))
-    return document_ids
