+++ sparkth/plugins/chat/routes/completions.py
@@ -27,10 +27,8 @@ from sparkth.plugins.chat.exceptions import RAGSearchError
 from sparkth.plugins.chat.lms_credentials import build_lms_credentials_message
 from sparkth.plugins.chat.messages import extract_query_text
 from sparkth.plugins.chat.prompt import REFUSAL_MESSAGE, get_learning_design_system_prompt
-from sparkth.plugins.chat.routes.utils import (
-    resolve_document_blocks,
-    resolve_tools,
-)
+from sparkth.plugins.chat.routes.utils import resolve_tools
+from sparkth.plugins.chat.routes.utils.rag_search import resolve_document_blocks
 from sparkth.plugins.chat.routes.utils.stream_processor import (
     ChatStreamProcessor,
     stream_out_of_scope_refusal,
