+++ sparkth/plugins/chat/routes/utils/stream_processor.py
@@ -23,9 +23,10 @@ from sparkth.lib.rag import (
     format_document_chunks_as_llm_context,
 )
 from sparkth.plugins.chat.constants import LLM_PROVIDER_API_ERRORS, RAG_CONTEXT_PROMPT
-from sparkth.plugins.chat.messages import collect_document_ids, extract_query_text
+from sparkth.plugins.chat.messages import extract_query_text
 from sparkth.plugins.chat.models import Conversation
 from sparkth.plugins.chat.prompt import REFUSAL_MESSAGE
+from sparkth.plugins.chat.routes.utils.rag_search import collect_document_ids
 from sparkth.plugins.chat.schemas import ChatMessage
 from sparkth.plugins.chat.service import ChatService
