+++ sparkth/plugins/chat/tests/test_rag_integration.py
@@ -14,10 +14,10 @@ from sparkth.lib.rag import (
     RetrievedChunk,
 )
 from sparkth.plugins.chat.messages import extract_query_text
-from sparkth.plugins.chat.routes.utils import resolve_document_blocks
+from sparkth.plugins.chat.routes.utils.rag_search import resolve_document_blocks
 from sparkth.plugins.chat.schemas import ChatMessage
 
-RETRIEVE_CONTEXT_PATH = "sparkth.plugins.chat.routes.utils.agentic_retrieve_context"
+RETRIEVE_CONTEXT_PATH = "sparkth.plugins.chat.routes.utils.rag_search.agentic_retrieve_context"
 
 
 def _user_msg(content: str | list[Any]) -> ChatMessage:
